fix: resolve circular dependency between types files
- Move VarsSchema and Vars type to separate vars.ts file
- Update imports to use new structure
- Fixes circular dependency: index.ts > redteam/types.ts > providers.ts
